refactor: extract shared test helpers into counts-mongo-helpers module
